% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mean.R
\name{my_mean}
\alias{my_mean}
\title{My own mean function}
\usage{
my_mean(x)
}
\arguments{
\item{x}{vector, of numbers}
}
\value{
numeric, the computed mean
}
\description{
My own mean function
}
\examples{

my_mean(1:10)
}
