% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stroop_df}
\alias{stroop_df}
\title{Example dataset from Rosenbaum, Mama, & Algom (2017)}
\format{A data frame with 200 rows and 4 variables:
\describe{
  \item{\code{subject}}{factor subject ids}
  \item{\code{congruency}}{factor congruency codes, Congruent vs. Incongruent}
  \item{\code{posture}}{factor whether the participant was sitting or standing}
  \item{\code{rts}}{numeric reaction time}
}}
\usage{
stroop_df
}
\description{
some sample reaction time data from a 2x2 Stroop experiment
}
\details{
Example dataset from Rosenbaum, Mama, & Algom (2017) downloaded from OSF, and loaded into R in long-format. See pre-processing script in data-raw.
}
\keyword{datasets}
